// packages/client-auto/src/types/birdeye/api/wallet.ts
export type WalletPortfolioItem = {
    address: string;
    name?: string;
    symbol: string;
    decimals: number;
    balance: string;
    uiAmount: number;
    chainId: string;
    logoURI: string;
    priceUsd: number;
    valueUsd: number;
};